add primitive gap detection

+++ simulation.py
@@ -14,12 +14,13 @@ class Simulation:
     def __init__(self, height: int = config['ENVIRONMENT_HEIGHT'], width: int = config['ENVIRONMENT_WIDTH']):
         self.height = height
         self.width = width
+        self.is_running = True
 
-        self.agent = Agent(position=np.array([0.5, 0.5]), velocity=np.array([0.1, 0.1]))
+        self.agent = Agent(position=np.array([0.5, 0.5]), velocity=np.array([0.0, 0.0]))
 
         self.obstacles = [Obstacle(position=np.array([-5.0, 3.0]), velocity=np.array([1.0, 0.2])),
                           Obstacle(position=np.array([5.0, 1.0]), velocity=np.array([-1.0, 0.4])),
-                          Obstacle(position=np.array([-5.0, -3.0]), velocity=np.array([0.0, 0.0])),
+                          Obstacle(position=np.array([-5.0, -3.0]), velocity=np.array([0.3, -0.3])),
                           Obstacle(position=np.array([3.0, -3.0]), velocity=np.array([0.2, 0.5]))]
 
         dpg.create_context()
@@ -27,24 +28,31 @@ class Simulation:
         viewport_height = math.floor(utils.meters_to_pixels(self.height))
         viewport_width = math.floor(utils.meters_to_pixels(self.width))
 
+        with dpg.handler_registry():
+            dpg.add_key_press_handler(callback=self.toggle_is_running)
+
         dpg.create_viewport(title="Dynamic Navigation Exploration", width=viewport_width, height=viewport_height)
         dpg.setup_dearpygui()
         dpg.show_viewport()
 
+    def toggle_is_running(self):
+        self.is_running = not self.is_running
+
     def run(self):
         while dpg.is_dearpygui_running():
-            if dpg.does_item_exist('canvas'):
-                dpg.delete_item('canvas')
+            if self.is_running:
+                if dpg.does_item_exist('canvas'):
+                    dpg.delete_item('canvas')
 
-            with dpg.viewport_drawlist(tag='canvas'):
-                for obstacle in self.obstacles:
-                    obstacle.update()
-                    obstacle.draw()
+                with dpg.viewport_drawlist(tag='canvas'):
+                    for obstacle in self.obstacles:
+                        obstacle.update()
+                        obstacle.draw()
 
-                self.agent.update(self.obstacles)
-                self.agent.draw()
+                    self.agent.update(self.obstacles)
+                    self.agent.draw()
 
-            time.sleep(config['TIMESTEP'])
+                time.sleep(config['TIMESTEP'])
             dpg.render_dearpygui_frame()
 
     def teardown(self):
